mailservice, message: Check file asynchronously and based on content type
When a file is opened we are guessing its type based on the file
extension, but this is likely not a good test as in unix a file can have
any extension.

So use instead a content-file based check. And do it asynchronously.

To avoid having two different APIs for file handling, since gio::File
already provides us everything we need, do not use a mix of rust native
and gio APIs for the initial handling, so that we do not have to depend
on tokio for further thread management while keeping all threaded.