message: Use async loading of the files contents
We were loading the files synchronously, while this is not a big deal
for small files it may lead to hangs in big ones.

At the same time we were doing unnecessary repeated IO actions that we
can avoid.

So just rely on gio to do file opening keeping all the code consistent