tests: Add test utility function to wait for futures to complete
Since commit 3f1fdbea98d we are using futures that rely on the glib main
loop, and so we needed to test them using the glib local spawn system

However, this was not something that worked, since `spawn_local` does
not wait for the futures to be completed (unless that's also awaited),
so we were basically not running the tests.

To fix this, add a test utility function to actually spawn a future,
and wait for it until it's completed.

To ensure this also works fine for tests, we need to catch any potential
panic, that may happen during the async execution.

As per this, replace all the tests to use this function and fix the
tests that were failing as per the recent changes